+++ deploy/mc-deploy/CHANGELOG.md
@@ -1,5 +1,8 @@
 ALSO: UPDATE version in pyproject.toml, add/push mc-deploy-X.Y.Z tag to github
 
+* 0.7.0: Docker work; settings_get_new takes args
+	Flavor tuple, as_login_user argument
+
 * 0.6.0: move deploy_cmd to base; rename deploy_helper to deploy_cmd_helper
 	move git checks from dokku to deploy_cmd_helper
 	have dburl and clone commands take instance id (prod/staging/USER)